SOGA-main/requirements.txt
antlr4-python3-runtime==4.10
numpy>=1.21.6
scipy>=1.7.3
sympy==1.9
scikit-learn==1.3.2
Mathics-omnibus==4.0.1
pandas==2.0.3
pymc3==3.11.5
psutil==5.9.8
Jinja2==3.1.4
natsort==8.4.0